rename constants file to appconstants due to module clash
