+++ src/appConstants.ts
