+++ src/routes/game/components/elements/playmat/Playmat.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useAppSelector } from 'app/Hooks';
 import { RootState } from 'app/Store';
-import { DEFAULT_PLAYMAT } from 'constants';
+import { DEFAULT_PLAYMAT } from 'appConstants';
 import styles from './Playmat.module.css';
 export const Playmat = ({ isPlayer }: { isPlayer: boolean }) => {
   let playmat = useAppSelector((state: RootState) =>
