+++ src/routes/game/components/opponentBoardGrid/OpponentBoardGrid.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './OpponentBoardGrid.module.css';
 import { useAppSelector } from 'app/Hooks';
 import { RootState } from 'app/Store';
-import { DEFAULT_PLAYMAT } from 'constants';
+import { DEFAULT_PLAYMAT } from 'appConstants';
 import HeadEqZone from '../zones/headEqZone/HeadEqZone';
 import PermanentsZone from '../zones/permanentsZone/PermanentsZone';
 import GraveyardZone from '../zones/graveyardZone/GraveyardZone';
