+++ src/routes/game/components/playerBoardGrid/PlayerBoardGrid.tsx
@@ -5,7 +5,7 @@ import BottomRow from '../bottomRow/BottomRow';
 import styles from './PlayerBoardGrid.module.css';
 import { useAppSelector } from 'app/Hooks';
 import { RootState } from 'app/Store';
-import { DEFAULT_PLAYMAT } from 'constants';
+import { DEFAULT_PLAYMAT } from 'appConstants';
 import HeadEqZone from '../zones/headEqZone/HeadEqZone';
 import PermanentsZone from '../zones/permanentsZone/PermanentsZone';
 import GraveyardZone from '../zones/graveyardZone/GraveyardZone';
